Getting rid of direct accessing getResourcePath for video
