+++ src/game/scripting/life.ts
@@ -5,7 +5,7 @@ import { SampleType } from '../data/sampleType';
 import { setMagicBallLevel } from '../../game/state';
 import { unimplemented } from './utils';
 import { WORLD_SCALE, getRandom } from '../../utils/lba';
-import { getResourcePath } from '../../resources';
+import { getVideoPath } from '../../resources';
 
 export const PALETTE = unimplemented();
 
@@ -378,7 +378,7 @@ export function PLAY_VIDEO(cmdState, video) {
         };
         this.game.setUiState({ skip: false,
             video: {
-                path: getResourcePath(`VIDEO_${video}`),
+                path: getVideoPath(video),
                 onEnded
             }});
         cmdState.skipListener = function skipListener() {
